refactor: CustomerService 인터페이스 삭제, 모든 의존관계 구현체로 변경
