+++ src/main/java/org/prgrms/voucherapplication/domain/voucher/controller/VoucherController.java
@@ -3,8 +3,8 @@ package org.prgrms.voucherapplication.domain.voucher.controller;
 import org.prgrms.voucherapplication.console.CommandType;
 import org.prgrms.voucherapplication.console.Input;
 import org.prgrms.voucherapplication.console.Output;
-import org.prgrms.voucherapplication.domain.customer.service.CustomerService;
 import org.prgrms.voucherapplication.domain.customer.controller.dto.ResponseBlacklist;
+import org.prgrms.voucherapplication.domain.customer.service.CustomerService;
 import org.prgrms.voucherapplication.domain.voucher.entity.Voucher;
 import org.prgrms.voucherapplication.domain.voucher.entity.VoucherConstructorException;
 import org.prgrms.voucherapplication.domain.voucher.entity.VoucherType;
